gbproxy: Send STATUS with original message back to the SGSN
Currently when patching is enabled and an error happens when
receiving a message from the SGSN, the patched message is sent back
with the PDU_IN_ERROR IE.

This patch modifies gbprox_rx_sig_from_sgsn() to copy the message
before it is patched, so that the original message can be used with
the STATUS message. gbprox_rx_data_from_sgsn() does all checks before
the message is patched, so copying is not necessary.

Sponsored-by: On-Waves ehf